Extract useAsyncCacheFetch and add test
