+++ cypress/support/component.js
@@ -20,3 +20,7 @@ import './commands'
 // require('./commands')
 // Example use:
 // cy.mount(<MyComponent />)
+
+// Import mount method from cypress/react
+import { mount } from 'cypress/react'
+Cypress.Commands.add('mount', mount)
